Correct Parsing Benchmark, remove RegexSet
The RegexSet option hasn't really been used since testing the intial
regexes in early releases.  Save the extra dereference in a few places,
this could lead to better performance on older Perls especially.

The benchmarks/01-parse.pl wasn't reporting messages per second.  It was
reporting how many blocks of "X" messages could be processed.  This
makes the output useless if we find another weird case we want in the
benchmarks.  Moved the benchmarks to a curried function and it now
reports message per second parsed.