Improve Key/Value Detection
Be smarter about handling key/value pairs.  This isn't perfect, but it's
a lot better than the previous revision.  This will strip all "brackets"
from the front and end of values and uses a smart regex to figure out
the boundaries for the values.  This allows for k/v detection on fields
like:

    dateStarted=2018-05-24 10:11:23 +00:00 dateStopped=2018-05-24 10:14:23 +00:00

And wind up with:

    {
        ...
        SDATA => {
            ...
            dateStarted => '2018-05-24 10:11:23 +00:00',
            dateStopped => '2018-05-24 10:14:23 +00:00',
            ...
        }
    }

Unstructured logs are fun!