Stop tests from failing without JSON::MaybeXS
Prior to this patch, tests were failing on the smoke test infrastructure
since `JSON::MaybeXS` is an optional dependency.

The tests now check for `JSON::MaybeXS` and only load and run the tests
if it's present.

Also adjusted the call of `decode_json()` to use the package path to
prevent issues if something else loads `JSON::MaybeXS` without importing
the `decode_json()` symbol.

Release as 4.7 to the CPAN.