Bugfix: IETF-syslog-proto23
This addresses an issue parsing IETF Syslog Protocol 23 formatted
messages. It adds a positive look-ahead to check if the preamble
"version" number is followed by a space. This differentiates the start
of an RFC-3339 date string from that of a version string. I.E. it
assumes, there is a space between the version number and date string.