Fix bugs in test regeneration
When running `t/bin/create_test_entry.pl --regenerate --auto`, there was
no processing of the `options` in the test file. This was due to my
horrible choice to use package variables as configuration options.
There's no easy way to "reset" them, so I punted on fixing it.

This new version of the script forks a child and then munges the
variables in the child. This automagically "resets" the global state.

Use the new code to regenerate the `t/data` YAML. This just sorted
keys, everything works as intended.